# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "JustoAI V2 API",
    "version": "2.0.0",
    "description": "API completa da plataforma JustoAI V2 para gestão jurídica com IA",
    "contact": {
      "name": "Equipe JustoAI",
      "email": "dev@justoai.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/api/ai/analyze": {
      "post": {
        "summary": "Análise com IA",
        "description": "Executa análise de documentos ou processos utilizando inteligência artificial",
        "tags": [
          "IA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "content"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "DOCUMENT",
                      "PROCESS",
                      "CASE"
                    ],
                    "description": "Tipo de análise"
                  },
                  "content": {
                    "type": "string",
                    "description": "Conteúdo a ser analisado (texto ou ID do documento)"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "deepAnalysis": {
                        "type": "boolean",
                        "default": false,
                        "description": "Realizar análise profunda (consome mais créditos)"
                      },
                      "language": {
                        "type": "string",
                        "default": "pt-BR"
                      },
                      "includeRecommendations": {
                        "type": "boolean",
                        "default": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Análise concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "analysisId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "summary": {
                          "type": "string"
                        },
                        "insights": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "recommendations": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "confidence": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 1,
                          "example": 0.95
                        },
                        "creditsConsumed": {
                          "type": "number",
                          "example": 10.5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "402": {
            "description": "Créditos insuficientes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}